apache-airflow/requirements.txt
alembic==1.7.4
anyio==3.3.4
apache-airflow==2.3.3
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apache-airflow-providers-ssh==3.1.0
apispec==3.3.2
argcomplete==1.12.3
asn1crypto==1.4.0
attrs==20.3.0
Babel==2.9.1
bcrypt==3.2.2
blinker==1.4
boto3==1.20.2
botocore==1.23.2
cachelib==0.9.0
cattrs==1.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
configparser==5.1.0
connexion==2.14.0
cron-descriptor==1.2.31
croniter==1.0.15
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
Flask==2.1.3
Flask-AppBuilder==4.1.2
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==4.4.2
Flask-Login==0.6.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
graphviz==0.18
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
iso8601==1.0.0
isodate==0.6.0
itsdangerous==2.1.2
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
markdown-it-py==2.1.0
MarkupSafe==1.1.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
oscrypto==1.2.1
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.8.0
psycopg2==2.9.1
py==1.11.0
pycparser==2.21
pycryptodomex==3.11.0
Pygments==2.10.0
PyJWT==2.4.0
pymongo==3.12.1
pymssql==2.2.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==7.1.2
pytest-mock==3.8.2
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
rich==12.5.1
s3transfer==0.5.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
snowflake-connector-python==2.7.0
SQLAlchemy==1.4.39
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
sshtunnel==0.4.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.3.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.7
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.6.0
